chore(build): move generated artifacts to ignored .generated dir — switch metadata refresh to prebuild
