+++ .gitignore
@@ -11,6 +11,7 @@
 # next.js
 /.next/
 /out/
+/.generated/
 
 # production
 /build
